Fixed problem in PACS server when receiving DICOM image without patient ID Added optional logo & header to drug prescription Added administration route to drug definition Added in-patient admission form template for Nyarugenge District Hospital Added out-patient consultation form template for Gabriel Touré university teaching hospital Update Burundi's clinical pathways logic (SPT) Minor cosmetic fixes to SPT form for uploading clinical pathways logic updates Possibility to generate paper prescriptions from existing electronic prescriptions Added real time pricing information to drugs dispensing screen for pharmacist

+++ src/be/openclinic/archiving/ScanDirectoryMonitor.java
@@ -542,7 +542,7 @@ public class ScanDirectoryMonitor implements Runnable{
 		int err=0;
     	DicomObject obj = Dicom.getDicomObjectNoPixels(file);
     	Debug.println("DICOM object: "+obj);
-    	if(obj!=null){
+    	if(obj!=null && obj.getString(Tag.PatientID)!=null){
         	Debug.println("Patient ID: "+obj.getString(Tag.PatientID));
         	int nPatientId = -1;
         	if(MedwanQuery.getInstance().getConfigInt("PACSUseNatregForPatientIdentification",0)==1) {
@@ -612,7 +612,16 @@ public class ScanDirectoryMonitor implements Runnable{
 		    		if(rs.next()){
 		    			Debug.println("Instance number "+sequence+" for study "+studyUid+" already exists");
 		    			err=0;
-		    			file.delete();
+		    			//Check if file also exists on disk
+		    			String sFileName = rs.getString("OC_PACS_FILENAME");
+		    			File f = new File(SCANDIR_BASE+"/"+SCANDIR_TO+"/"+sFileName);
+		    			if(!f.exists()) {
+		    				//The file is missing. Move the received file to the destination file
+		    				Debug.println(moveFile(file,f));
+		    			}
+		    			else {
+		    				file.delete();
+		    			}
 		    		}
 		    		else {
 		    			//file does not exist yet, insert it
